添加接口和注释说明; 修正部分方法中商品ID参数类型错误的问题; 查找商品详情添加了是否被收藏; 添加了收藏商品方法;

+++ src/main/java/com/orangeSoft/market/service/ICommodityService.java
@@ -34,8 +34,9 @@ public interface ICommodityService extends IService<Commodity> {
 
     /**
      * 根据商品ID查找商品详情
+     *
      * @param commodityId 商品ID
      * @return Result
      */
-    Result.JSONResultMap getCommodityById(Integer commodityId);
+    Result.JSONResultMap getCommodityById(Long commodityId);
 }
